+++ src/lib/state_files.h
@@ -11,8 +11,8 @@
 #ifndef MYMPD_STATE_FILES_H
 #define MYMPD_STATE_FILES_H
 
-#include "dist/libmympdclient/include/mpd/client.h"
 #include "dist/sds/sds.h"
+#include "src/lib/mpdclient.h"
 #include "src/lib/validate.h"
 
 #include <stdbool.h>